Refactoring favorites.ui vers - favorites-actions.js - favorites-init.js - favorite-popup-details.js - favorite-popup-dialogs.js - favorite popup.js - favorite-rendering.js - Suprresion de favorite-ui.js
