+++ assets/js/components/favorites/favoritesModule.js
@@ -1,8 +1,11 @@
-// * favoritesModule.js - Point d'entrée navigateur pour le module favoris (initialisation automatique)
-import { FavoritesUI } from './favorite-ui.js';
+// favoritesModule.js - Point d'entrée navigateur pour le module favoris (initialisation automatique)
+import { FavoritesUI } from './favoritesindex.js';
 
 if (typeof window !== 'undefined') {
+    // Exposer l'interface unifiée au niveau global
     window.FavoritesUI = FavoritesUI;
+    
+    // Initialiser automatiquement au chargement du DOM
     if (document.readyState === 'loading') {
         document.addEventListener('DOMContentLoaded', () => FavoritesUI.init());
     } else {
@@ -10,6 +13,7 @@ if (typeof window !== 'undefined') {
     }
 }
 
+// Support pour l'exportation CommonJS si nécessaire
 if (typeof module !== 'undefined' && module.exports) {
     module.exports = FavoritesUI;
 }
